Fixed potential bug: FLOAT types where read from database using ResultSet.getFloat() which can lead to truncation, so now we use getDouble(). Likewise, when writing results, the output tables are created with DECIMAL(6,2). The output tables are now FLOAT so that in most databases, it uses more precision. But we still want a way to customize this (for instance, in MS SQL Server, FLOAT is 8 bytes double while in MySQL it is 4 bytes single precision float)

+++ src/main/java/com/ibm/opl/customdatasource/JdbcCustomDataSource.java
@@ -113,7 +113,7 @@ public class JdbcCustomDataSource extends IloCustomOplDataSource {
                 if (leaf == Type.INTEGER)
                     handler.addIntItem(rs.getInt(1));
                 else if (leaf == Type.FLOAT)
-                    handler.addNumItem(rs.getFloat(1));
+                    handler.addNumItem(rs.getDouble(1));
                 else if (leaf == Type.STRING)
                     handler.addStringItem(rs.getString(1));
             }
@@ -155,7 +155,7 @@ public class JdbcCustomDataSource extends IloCustomOplDataSource {
                     if (oplFieldsType[column] == Type.INTEGER) {
                         handler.addIntItem(rs.getInt(columnName));
                     } else if (oplFieldsType[column] == Type.FLOAT) {
-                        handler.addNumItem(rs.getFloat(columnName));
+                        handler.addNumItem(rs.getDouble(columnName));
                     } else if (oplFieldsType[column] == Type.STRING) {
                         handler.addStringItem(rs.getString(columnName));
                     }
